Add files via upload
A wider education on Data Science tools, such as using python packages like pandas and numpy, SQL, Tableau, machine learning and regression analysis. Culminating in a Capstone project that analysed the data of property sales in Ames, Iowa between 2006 and 2010, and drawing useful, clear insights from the data by creating training and testing sets to run models and produce visual results for expected changes in sales prices based on a host of variables.